requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
mysqlclient==2.2.7      # For MySQL
pandas==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
yfinance==0.2.33
pytz==2023.3
python-dateutil==2.8.2
gunicorn==21.2.0        # For production
whitenoise==6.6.0       # For static files